+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 2.3.2
 
+  - move check rule toolchain code
+
 ## 2.3.1
 
   - make doxygen install as optional