+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 2.3.2
 
+  - use web request instead bits transfert in powershell toolchain
   - move check project name toolchain code
   - move check rule toolchain code